+++ src/locales/fi.json
@@ -146,6 +146,7 @@
       "section": "Osasto",
       "selection": "Valinta",
       "serviceMap": "Palvelukartta",
+      "stickerNumber": "Tarranumero",
       "storageOnIce": "Jäissä säilytys",
       "summerStorageForBoats": "Veneiden kesäsäilytys",
       "summerStorageForDockingEquipment": "Telakointi\u00ADtarvikkeiden kesäsäilytys",
